requirements.txt
aiogram==2.25.2
openai>=1.0.0
python-dotenv
tradingview_ta
mplfinance
matplotlib
pandas
aiohttp>=3.8.0,<3.9.0
pandas_ta
